Crate edn_rs

Re-exports

pub use edn::Error as EdnError;

pub use edn::Double;

pub use edn::Edn;

pub use edn::List;

pub use edn::Map;

pub use edn::Set;

pub use edn::Vector;

pub use serialize::Serialize;

Modules

edn

Edn type implementation

serialize

Serialization module for most possible types. Tuples are limited between (A, B) and (A, B, C, D, E, F), any other tuple needs to be implemented by the trait Serialize. This module requires #[macro_use] for structs.

Macros

edn

Macro to parse EDN into Rust Spec:

hmap

Creates a HashMap from a seq of $key => $value, hmap!{a => "b", c => "d"}

hset

Creates a HashSet from a seq of $x, set!{1, 2, 3, 4}

map

Creates a BTreeMap from a seq of $key => $value, map!{a => "b", c => "d"}

set

Creates a BTreeSet from a seq of $x, set!{1, 2, 3, 4}

Traits

Deserialize

public trait to be used to Deserialize structs.

Functions

from_edn

from_edn deserializes an EDN type into a T type that implements Deserialize. Response is Result<T, EdnError>

from_str

from_str deserializes an EDN String into type T that implements Deserialize. Response is Result<T, EdnError>

to_string

Function for converting Rust types into EDN Strings. For it to work, the type must implement the Serialize trait. Use #[derive(Serialize)] from edn-derive crate.